% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-edina.R
\name{autoplot.auto_edina}
\alias{autoplot.auto_edina}
\title{Graph the Auto EDINA Object}
\usage{
\method{autoplot}{auto_edina}(
  object,
  type = c("selection", "guessing", "slipping", "evolution"),
  ...
)
}
\arguments{
\item{object}{An \code{auto_edina} object.}

\item{type}{Kind of graph to display. Valid types: \code{"selection"} or \code{"evolution"}.}

\item{...}{Not used.}
}
\value{
A \code{ggplot2} object.
}
\description{
Presents either the fitting of model heuristics or the evolution of parameters
on a graph
}
