% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-selection-criteria.R
\name{BIC.edina}
\alias{BIC.edina}
\title{Calculate BIC for EDINA models.}
\usage{
\method{BIC}{edina}(object, ...)
}
\arguments{
\item{object}{An \code{edina} object}

\item{...}{Not used.}
}
\description{
Calculate BIC for EDINA models.
}
