% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{edina}
\alias{edina}
\title{EDINA estimation routine}
\usage{
edina(data, k = 3, burnin = 10000, chain_length = 20000)
}
\arguments{
\item{data}{Binary responses to assessements in \code{matrix}
form with dimensions \eqn{N \times J}{N x J}.}

\item{k}{Number of Attribute Levels as a positive \code{integer}.}

\item{burnin}{Number of Observations to discard on the chain.}

\item{chain_length}{Length of the MCMC chain}
}
\value{
An \code{edina} object that contains:
\itemize{
\item \code{coefficients}: Estimated coefficients of the model fit
\item \code{loglike_summed}: Summed log-likelihood
\item \code{loglike_pmean}: Meaned version of log-likelihood
\item \code{pi_classes}: Latent classes
\item \code{avg_q}: Estimated Averaged Q Matrix
\item \code{est_q}: Estimated Dichotomous Q Matrix
\item \code{or_tested}: Odds Ratio used in the model selection.
\item \code{sample_or}: Odds Ratio for the sample.
\item \code{n}: Number of Observations
\item \code{j}: Number of Items
\item \code{k}: Number of Traits
\item \code{burnin}: Amount of iterations to discard
\item \code{chain_length}: Amount of iterations to retain.
\item \code{timing}: Duration of the run
\item \code{dataset_name}: Name of the data set used in estimation.
}
}
\description{
Exploratory Determinatistic Input, Noise and Gate Model (EDINA)
}
\examples{
\dontrun{
library("tmsadata")

# Load data
data("trial_matrix", package="tmsadata")

# Coerce to matrix
trial_matrix = as.matrix(trial_matrix)

edina_model = edina(trial_matrix, k = 2)
}

}
