% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-auto-edina.R
\name{model_selection_graph}
\alias{model_selection_graph}
\title{View Model Selection Statistics Across Models}
\usage{
model_selection_graph(x, ...)
}
\arguments{
\item{x}{An \code{auto_edina} or \code{auto_errum} object.}

\item{...}{Not used}
}
\value{
A \code{ggplot2} object
}
\description{
Displays information about the value of each model information criterion
for a given model across the dimensions the Q matrix is estimated.
}
\seealso{
\code{\link[=autoplot.auto_edina]{autoplot.auto_edina()}}
}
