% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{new_edina}
\alias{new_edina}
\title{Construct an EDINA object}
\usage{
new_edina(
  coefs,
  loglike_summed,
  loglike_pmean,
  pis,
  avg_q,
  est_q,
  or_tested,
  sample_or,
  n,
  burnin,
  chain_length,
  timing,
  dataset_name
)
}
\arguments{
\item{coefs}{Matrix with means of guessing and slipping coefficients.}

\item{loglike_summed}{Log-likelihood over the iterations.}

\item{loglike_pmean}{Log-likelihood at the mean point}

\item{pis}{Estimated latent classes}

\item{avg_q}{Averaged Q Matrix over Iterations}

\item{est_q}{Estimated Q Matrix}

\item{or_tested}{Sample empirical odds ratio compared against simulated
odds ratio divided by the number of simulations.}

\item{sample_or}{Sample empirical odds ratio based on the trial matrix.}

\item{n}{Number of Observations}

\item{burnin}{Number of Observations to discard on the chain.}

\item{chain_length}{Length of the MCMC chain}

\item{timing}{Number of Seconds that have elapsed since run time.}

\item{dataset_name}{Name of the data set the estimation procedure ran on.}
}
\description{
Constructor function to satisfy the EDINA object definition
}
\keyword{internal}
