% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{new_edina_summary}
\alias{new_edina_summary}
\title{Construct a Summary EDINA object}
\usage{
new_edina_summary(edina, model_fit, alpha)
}
\arguments{
\item{edina}{An \code{edina} object}

\item{model_fit}{Computed model heuristic value}

\item{alpha}{The region used in the computation of the heuristic.}
}
\description{
Constructor function to satisfy the Summary EDINA object definition
}
\keyword{internal}
