% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-auto-edina.R
\name{parameter_evolution_graph}
\alias{parameter_evolution_graph}
\title{View Slipping and Guessing Parameter Changes Across Models}
\usage{
parameter_evolution_graph(x, ...)
}
\arguments{
\item{x}{An \code{auto_edina} or \code{auto_errum} object.}

\item{...}{Not used}
}
\description{
Displays the slipping and guessing parameter changes for each model across
the dimensions the Q matrix is estimated.
}
