% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-selection-criteria.R
\name{PPP}
\alias{PPP}
\title{Model Heuristic used for Model Selection}
\usage{
PPP(object, ...)
}
\arguments{
\item{object}{An \code{edina} object}

\item{...}{Not used.}
}
\value{
A \code{double} that is the mean.
}
\description{
Computes the model heuristic for model selection.
}
