% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-edina.R
\name{print.auto_edina}
\alias{print.auto_edina}
\title{Print method for \code{auto_dina}}
\usage{
\method{print}{auto_edina}(x, ...)
}
\arguments{
\item{x}{An \code{auto_edina} object}

\item{...}{Additional methods passed onto the \code{print.data.frame} method.}
}
\description{
Custom print method for displaying the results of the Auto EDINA method.
}
