% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{print.edina}
\alias{print.edina}
\title{Printing out the EDINA Object}
\usage{
\method{print}{edina}(x, binary_q = FALSE, ...)
}
\arguments{
\item{x}{An \code{edina} object}

\item{binary_q}{Boolean to indicate whether the \emph{Q} matrix is shown in
dichotomous form or in an estimated form.}

\item{...}{Additional methods passed onto the \code{print.matrix} method.}
}
\description{
Custom print method for computing the EDINA.
}
