% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-edina.R
\name{print.summary_auto_edina}
\alias{print.summary_auto_edina}
\title{Print the \code{auto_edina} model summary}
\usage{
\method{print}{summary_auto_edina}(x, ...)
}
\arguments{
\item{x}{A \code{summay_auto_edina} object}

\item{...}{Additional values passed onto the \code{print.data.frame} method.}
}
\value{
None.

The function provides a side-effect of displaying the overview of
computed results across all models estimated.
}
\description{
Custom method for displaying the results of the \code{summary(auto_edina)}.
}
