% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{print.summary_edina}
\alias{print.summary_edina}
\title{Printing out the Summary EDINA Object}
\usage{
\method{print}{summary_edina}(x, binary = FALSE, ...)
}
\arguments{
\item{x}{A \code{summary_edina} object}

\item{binary}{Boolean to indicate whether the \emph{Q} matrix is shown in
dichotomous form or in an estimated form.}

\item{...}{Past onto the \code{print.data.frame} method.}
}
\value{
None.

The function provides a side-effect of displaying the overview of
the model estimated.
}
\description{
Custom print method for displaying the EDINA model summary information.
}
