% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{q_graph}
\alias{q_graph}
\alias{q_graph.auto_edina}
\alias{q_graph.edina}
\alias{q_graph.matrix}
\alias{q_graph.q_matrix}
\title{Graph Q Matrix}
\usage{
q_graph(x, ...)

\method{q_graph}{auto_edina}(x, binary_q = TRUE, ic = c("heuristic", "bic", "dic"), ...)

\method{q_graph}{edina}(x, binary_q = TRUE, ...)

\method{q_graph}{matrix}(x, ...)

\method{q_graph}{q_matrix}(x, ...)
}
\arguments{
\item{x}{Either an \code{edina}, \code{errum}, or \code{q_matrix} object.}

\item{...}{Additional paramters not used}

\item{binary_q}{Classified Q matrix or a rounded Q matrix.}

\item{ic}{Selection criteria}
}
\description{
Provides a heatmap approach to showing the estimated binary or averaged
values of the Q Matrix.
}
