% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{q_matrix}
\alias{q_matrix}
\title{Create a Q Matrix Object}
\usage{
q_matrix(x)
}
\arguments{
\item{x}{Either a \code{data.frame} or \code{matrix}.}
}
\description{
Provides a way to cast an object as a \code{"q_matrix"}.
}
