% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-edina.R
\name{summary.auto_edina}
\alias{summary.auto_edina}
\title{Summarize \code{auto_edina} model data}
\usage{
\method{summary}{auto_edina}(object, ...)
}
\arguments{
\item{object}{An \code{auto_edina} object}

\item{...}{Additional methods passed onto the \code{print.data.frame} method.}
}
\value{
An invisible \code{data.frame} containing criterion information.
}
\description{
Custom method for displaying the results of the \code{auto_edina}.
}
