% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{summary.edina}
\alias{summary.edina}
\title{Summarizing the EDINA Object}
\usage{
\method{summary}{edina}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{An \code{edina} object}

\item{alpha}{Defining region to indicate the level of extremeness
the data must before the model is problematic.}

\item{...}{Not used.}
}
\description{
Determine whether the \code{edina} object is summarized appropriately.
}
