% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{as_q_matrix}
\alias{as_q_matrix}
\alias{as_q_matrix.data.frame}
\alias{as_q_matrix.matrix}
\alias{as_q_matrix.default}
\title{Coerce \code{data.frame} and \code{matrix} classes to Q Matrix.}
\usage{
as_q_matrix(x, ...)

\method{as_q_matrix}{data.frame}(x, ...)

\method{as_q_matrix}{matrix}(x, ...)

\method{as_q_matrix}{default}(x, ...)
}
\arguments{
\item{x}{Either a \code{data.frame} or \code{matrix}.}

\item{...}{Not used}
}
\description{
\code{as.q_matrix} acts as an aliases.
}
