% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-edina.R
\name{auto_edina}
\alias{auto_edina}
\title{Auto EDINA model selection routine}
\usage{
auto_edina(
  data,
  k = 2:4,
  burnin = 10000,
  chain_length = 20000,
  save_results = FALSE,
  save_filename = "edina_model_data"
)
}
\arguments{
\item{data}{Binary responses to assessements in \code{matrix}
form with dimensions \eqn{N \times J}{N x J}.}

\item{k}{Number of Attribute Levels as a positive \code{integer}.}

\item{burnin}{Number of Observations to discard on the chain.}

\item{chain_length}{Length of the MCMC chain}

\item{save_results}{Output intermediary estimations to R binary.
Default: \code{FALSE}}

\item{save_filename}{Name stem to save each \code{k} iteration to. Relevant if
\code{save_results} is enabled.}
}
\value{
An \code{auto_edina} object
}
\description{
Automatically select an appropriate $K$ dimension for a $Q$ matrix
under the Exploratory Determinatistic Input, Noise And gate (EDINA) Model.
}
\examples{
\dontrun{
library("tmsadata")

# Load data
data("trial_matrix", package="tmsadata")

# Coerce to matrix
trial_matrix = as.matrix(trial_matrix)

edina_models = auto_edina(trial_matrix, k = 1:2)
}
}
