% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-selection.R
\name{best_model}
\alias{best_model}
\alias{best_model.auto_edina}
\title{Extract the Best Model}
\usage{
best_model(x, ...)

\method{best_model}{auto_edina}(x, ic = c("ppp", "bic", "dic"), ...)
}
\arguments{
\item{x}{An \code{auto_edina} object}

\item{...}{Not used.}

\item{ic}{Information criterion name. Default \code{"ppp"}.}
}
\value{
An \code{edina} model object corresponding to the smallest value of requested
information criterion.
}
\description{
Extracts the best model from the \verb{auto_*()} search procedure.
}
\seealso{
\code{\link[=DIC.edina]{DIC.edina()}},
\code{\link[=BIC.edina]{BIC.edina()}},
\code{\link[=PPP.edina]{PPP.edina()}}
}
