% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{best_model}
\alias{best_model}
\title{Extract the Best Model}
\usage{
best_model(x, ...)
}
\arguments{
\item{x}{An \code{auto_edina} object}

\item{...}{Not used.}

\item{ic}{Selection criteria}
}
\value{
An \code{edina} object
}
\description{
Extracts the best model from the \verb{auto_*()} search procedure.
}
