% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-selection.R
\name{BIC.edina}
\alias{BIC.edina}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{BIC}{edina}(object, ...)
}
\arguments{
\item{object}{An \code{edina} object}

\item{...}{Not used.}
}
\value{
The BIC value of the given model.
}
\description{
Calculate BIC for EDINA models.
}
\section{BIC Computation Procedure}{


\eqn{BIC = -2 \log p\left( {\mathbf{y}| \mathbf{\hat{\theta}} } \right) + ((k+2)*j + 2^k)\log(n)}
}

\seealso{
\code{\link[=PPP.edina]{PPP.edina()}}, \code{\link[=DIC.edina]{DIC.edina()}}

\code{\link[=PPP.edina]{PPP.edina()}}, \code{\link[=DIC.edina]{DIC.edina()}}
}
