% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-selection.R
\name{DIC}
\alias{DIC}
\alias{DIC.edina}
\title{Deviance Information Criterion (DIC)}
\usage{
DIC(object, ...)

\method{DIC}{edina}(object, ...)
}
\arguments{
\item{object}{An \code{edina} object}

\item{...}{Not used.}
}
\value{
The DIC value of the given model.
}
\description{
Calculate DIC for EDINA models.
}
\section{DIC Computation Procedure}{


\eqn{DIC = -2\left({\log p\left( {\mathbf{y}| \mathbf{\hat{\theta}} } \right)  - 2\left( {\log p\left( {\mathbf{y}| \mathbf{\hat{\theta}} } \right) - \frac{1}{N}\sum\limits_{n = 1}^N {\log p\left( {\mathbf{y}|{\mathbf{\theta} _s}} \right)} } \right)} \right)}
}

\seealso{
\code{\link[=PPP.edina]{PPP.edina()}}, \code{\link[=BIC.edina]{BIC.edina()}}
}
