% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{DIC}
\alias{DIC}
\alias{DIC.edina}
\title{Compute the Deviance Information Criterion (DIC)}
\usage{
DIC(object, ...)

\method{DIC}{edina}(object, ...)
}
\arguments{
\item{object}{An \code{edina} object}

\item{...}{Not used.}
}
\description{
Calculate DIC for EDINA models.
}
