% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{PPP}
\alias{PPP}
\title{Model Heuristic used for Model Selection}
\usage{
PPP(object, alpha = 0.05)
}
\arguments{
\item{object}{An \code{edina} object}

\item{alpha}{Defining region to indicate the level of extremeness
the data must before the model is problematic.}
}
\value{
A \code{double} that is the mean.
}
\description{
Computes the model heuristic for model selection.
}
