% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-edina.R
\name{print.auto_edina}
\alias{print.auto_edina}
\title{Print method for \code{auto_edina}}
\usage{
\method{print}{auto_edina}(x, ...)
}
\arguments{
\item{x}{An \code{auto_edina} object}

\item{...}{Additional values passed onto the \code{print.data.frame} method.}
}
\value{
None.

The function provides a side-effect of displaying the overview of
computed results across all models estimated.
}
\description{
Custom print method for displaying the results of the Auto EDINA method.
}
