% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-matrix.R
\name{print.q_matrix}
\alias{print.q_matrix}
\title{Printing out a Q Matrix Object}
\usage{
\method{print}{q_matrix}(x, ...)
}
\arguments{
\item{x}{An \code{q_matrix} object}

\item{...}{Additional methods passed onto the \code{print.matrix} method.}
}
\value{
An invisible \code{matrix} without the \code{q_matrix} class displayed as a part
of the output displayed.
}
\description{
Custom print method for the Q Matrix Object.
}
\examples{
# Q matrix values
x = matrix(c(1, 0, 0, 1), nrow = 2)

# Show Q matrix structure
q_matrix(x)
}
\seealso{
\code{\link[=q_matrix]{q_matrix()}}, \code{\link[=as_q_matrix]{as_q_matrix()}}
}
